src/envs/finpulse_env/server/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
numpy==1.24.3
requests==2.31.0
pydantic==2.5.0
alpaca-py==0.43.2
python-dotenv==1.0.0
